Modify IB flex query importer
With the relocation of the structure of the IB flex query importer
we adjust the structure and the importer gets its own directory similar to the PDF and CSV importers.
