+++ name.abuchen.portfolio/META-INF/MANIFEST.MF
@@ -8,6 +8,7 @@ Export-Package: name.abuchen.portfolio.checks,
  name.abuchen.portfolio.datatransfer,
  name.abuchen.portfolio.datatransfer.actions,
  name.abuchen.portfolio.datatransfer.csv,
+ name.abuchen.portfolio.datatransfer.ibflex,
  name.abuchen.portfolio.datatransfer.pdf,
  name.abuchen.portfolio.events,
  name.abuchen.portfolio.json,
